Get version info from runtime/debug
Use the new runtime/debug package to get to version info. This is useful
for when someone makes a build locally as the date, revision and
modified state now become visible.

The version, commit and date default to "unknown" but are still
explicitly set during release by goreleaser using ldflags. This ensures
release builds are reproducible and don't require a git checkout to be
present.